requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

dishka==1.5.2
openai==1.72.0
openapi-spec-validator==0.7.1
python-dotenv==1.1.0
